fix(readiness): cover lifecycle-only install surfaces
Work-Item: CodySwannGT/lisa#1903
