+++ src/core/upstream-evidence-manifest.ts
@@ -7386,6 +7386,7 @@ export const UPSTREAM_SURFACE_MANIFEST: Readonly<Record<string, true>> =
     "src/cli/doctor-readiness-operations.ts": true,
     "src/cli/doctor-readiness-release-path.ts": true,
     "src/cli/doctor-readiness-shared.ts": true,
+    "src/cli/doctor-readiness-supply-chain-evidence.ts": true,
     "src/cli/doctor-readiness-supply-chain-scan.ts": true,
     "src/cli/doctor-readiness-supply-chain.ts": true,
     "src/cli/doctor-readiness-types.ts": true,
